[feat] 아파트 디테일 정보 타입 정리 #37
