+++ src/components/Main/ApartInfo/ApartMainInfo/ApartBuildInfo/ApartBuildInfoList.tsx
@@ -1,5 +1,5 @@
 import styles from "./ApartBuildInfoList.module.scss";
-import { TApartBuildInfoList } from "@/types/TMain/TApartBuildInfoList";
+import { TApartBuildInfoList } from "@/types/TMain/TApartBuildInfoListTypes";
 
 export default function ApartBuildInfoList(props: TApartBuildInfoList) {
   return (
