+++ src/types/TMain/TCurrentApartsStatusListTypes.ts
