+++ src/types/TMain/TNoriceInfoTypes.ts
